Adding small changes.  To note, as the radial distance increases there is less agreement between Oregon version and threaded version. I believe this has to do with no lock being held as the sensor array is updated in the Boost version.  Possible solution is to create a local sensor array for each thread and at the end of execution copy to the medium's (global) sensor array after grabbing a lock, ensuring  all values are updated properly.

+++ medium.cpp
@@ -56,7 +56,7 @@ void Medium::absorbEnergy(const double z, const double energy)
 #endif
 	
 	double r = fabs(z);
-	short ir = (short)(r/radial_bin_size);
+	int ir = (r/radial_bin_size);
 	if (ir >= num_radial_pos) {
 		ir = num_radial_pos;
 	}
